refactor: service 패키지 삭제
