+++ src/main/java/lotto/helper/ValidInputHelper.java
@@ -1,4 +1,4 @@
-package lotto.service.helper;
+package lotto.helper;
 
 import java.util.function.Consumer;
 import java.util.function.Supplier;
